Move structs into respective named files
